libipa: awb_bayes: Remove broken handling for lux == 0
Do not attempt to construct a picecewise linear function from one point
if the lux value is 0. The calibration data usually contains an entry
for a lux value of 0, so no special handling should be needed for this
case. This prevents an out-of-bounds access when using the AwbBayes
algorithm in low-light conditions.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>